refacto(ActivitySearchFilter): Move mobile version to a separate file
