+++ frontend/src/components/ActivitySearchFilter/ActivitySearchFilter.tsx
@@ -1,109 +1,14 @@
-import { Arrow } from 'components/Icons/Arrow';
 import { ChevronDown } from 'components/Icons/ChevronDown';
 import { MoreHorizontal } from 'components/Icons/MoreHorizontal';
 import { Link } from 'components/Link';
 import { Activity } from 'modules/activities/interface';
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
-import Select from 'react-select';
 import { routes } from 'services/routes';
-import { colorPalette, getSpacing, shadow } from 'stylesheet';
 
 import { ActivityButton } from './ActivityButton';
 import { useActivitySearchFilter } from './useActivitySearchFilter';
-import { useActivitySearchFilterMobile } from './useActivitySearchFilterMobile';
-
-const adaptActivityForSelect = (activity: Activity): { value: string; label: string } => ({
-  value: `${activity.id}`,
-  label: activity.name,
-});
-
-const ActivitySearchFilterMobile: React.FC<{ className?: string; activities: Activity[] }> = ({
-  className,
-  activities,
-}) => {
-  const { selectedActivity, updateSelectedActivity } = useActivitySearchFilterMobile();
-
-  return (
-    <div className={`${className ?? ''} flex space-x-4 items-center`}>
-      <Select
-        className="flex-1"
-        options={activities.map(adaptActivityForSelect)}
-        styles={selectStyles}
-        isSearchable={false}
-        placeholder={<FormattedMessage id="home.selectPlaceholder" />}
-        onChange={activity => updateSelectedActivity(activity?.value ?? null)}
-      />
-      {selectedActivity !== null ? (
-        <Link href={`${routes.SEARCH}?activity=${selectedActivity}`}>
-          <ValidateButton />
-        </Link>
-      ) : (
-        <ValidateButton />
-      )}
-    </div>
-  );
-};
-
-const ValidateButton: React.FC = () => (
-  <div className="bg-primary1 hover:bg-primary1-light shadow-lg text-white rounded-lg p-3.5 cursor-pointer transition-all">
-    <Arrow size={24} />
-  </div>
-);
-
-const selectStyles = {
-  control: (styles: any) => ({
-    ...styles,
-    backgroundColor: 'white',
-    boxShadow: shadow.large,
-    border: 'none',
-    padding: getSpacing(2),
-    borderRadius: getSpacing(2),
-    flexGrow: 1,
-    ':focus': {
-      border: 'none',
-    },
-  }),
-  valueContainer: (styles: any) => ({
-    ...styles,
-    padding: 'none',
-    overflow: 'visible',
-  }),
-  singleValue: (styles: any) => ({
-    ...styles,
-    color: colorPalette.greyDarkColored,
-    paddingLeft: getSpacing(2),
-  }),
-  menu: (styles: any) => ({
-    ...styles,
-    padding: `${getSpacing(2)} 0`,
-    margin: 0,
-    border: 'none',
-    boxShadow: shadow.large,
-    borderRadius: getSpacing(2),
-  }),
-  option: (styles: any) => ({
-    ...styles,
-    backgroundColor: 'white',
-    paddingLeft: getSpacing(4),
-    color: colorPalette.greyDarkColored,
-    ':hover': {
-      backgroundColor: colorPalette.primary2,
-    },
-  }),
-  indicatorSeparator: (styles: any) => ({
-    color: 'transparent',
-  }),
-  placeholder: (styles: any) => ({
-    ...styles,
-    color: colorPalette.greyDarkColored,
-    paddingLeft: '8px',
-    ':focus': {
-      outline: 'none',
-      color: 'white',
-    },
-  }),
-};
+import { ActivitySearchFilterMobile } from './ActivitySearchFilterMobile';
 
 interface Props {
   className?: string;
