+++ frontend/src/components/ActivitySearchFilter/ActivitySearchFilterMobile/useActivitySearchFilterMobile.tsx
